add policy_rule_id to session event for easier debugging of policy manager rules (Comment #12644)
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/hades/src@42998 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ src/com/untangle/node/ipsec_vpn/IpsecVpnApp.java
@@ -46,7 +46,7 @@ public class IpsecVpnApp extends NodeBase
 
     private static final Logger logger = Logger.getLogger(IpsecVpnApp.class);
     private final VirtualUserTable virtualUserTable = new VirtualUserTable();
-    private final Long policyId = getNodeSettings().getPolicyId();
+    private final Integer policyId = getNodeSettings().getPolicyId();
     private final PipelineConnector[] connectors = new PipelineConnector[0];
     private final IpsecVpnManager manager = new IpsecVpnManager();
